refactor & unit tests
