+++ query/src/storages/fuse/io/write/block_stream_writer.rs
@@ -32,7 +32,7 @@ use crate::pipelines::processors::transforms::ExpressionExecutor;
 use crate::sessions::TableContext;
 use crate::storages::fuse::io::TableMetaLocationGenerator;
 use crate::storages::fuse::operations::column_metas;
-use crate::storages::fuse::statistics::accumulator::BlockStatistics;
+use crate::storages::fuse::statistics::BlockStatistics;
 use crate::storages::fuse::statistics::StatisticsAccumulator;
 use crate::storages::index::ClusterKeyInfo;
