+++ query/src/storages/fuse/io/write/block_writer.rs
@@ -26,7 +26,7 @@ use crate::storages::fuse::io::retry;
 use crate::storages::fuse::io::retry::Retryable;
 use crate::storages::fuse::io::TableMetaLocationGenerator;
 use crate::storages::fuse::operations::util;
-use crate::storages::fuse::statistics::accumulator;
+use crate::storages::fuse::statistics::gen_columns_statistics;
 
 pub struct BlockWriter<'a> {
     location_generator: &'a TableMetaLocationGenerator,
@@ -48,7 +48,7 @@ impl<'a> BlockWriter<'a> {
         let data_accessor = &self.data_accessor;
         let row_count = block.num_rows() as u64;
         let block_size = block.memory_size() as u64;
-        let col_stats = accumulator::columns_statistics(&block)?;
+        let col_stats = gen_columns_statistics(&block)?;
         let (file_size, file_meta_data) = write_block(block, data_accessor, &location).await?;
         let col_metas = util::column_metas(&file_meta_data)?;
         let cluster_stats = None; // TODO confirm this with zhyass
