+++ query/src/storages/fuse/statistics/accumulator.rs
@@ -17,9 +17,7 @@ use std::collections::HashMap;
 
 use common_arrow::parquet::metadata::ThriftFileMetaData;
 use common_datablocks::DataBlock;
-use common_datavalues::prelude::*;
 use common_exception::Result;
-use common_functions::aggregates::eval_aggr;
 use common_fuse_meta::meta::BlockMeta;
 use common_fuse_meta::meta::ClusterStatistics;
 use common_fuse_meta::meta::ColumnId;
@@ -29,12 +27,8 @@ use common_fuse_meta::meta::StatisticsOfColumns;
 use common_fuse_meta::meta::Versioned;
 
 use crate::storages::fuse::operations::column_metas;
-use crate::storages::index::MinMaxIndex;
-use crate::storages::index::SupportedType;
-
-// pub trait FieldEnumerator {
-//    fn fields(&self) -> dyn Iterator<iter = (u32, &DataField)>;
-//}
+use crate::storages::fuse::statistics::block_statistics::BlockStatistics;
+use crate::storages::fuse::statistics::column_statistic;
 
 #[derive(Default)]
 pub struct StatisticsAccumulator {
@@ -62,7 +56,7 @@ impl StatisticsAccumulator {
         self.summary_block_count += 1;
         self.summary_row_count += row_count;
         self.in_memory_size += block_in_memory_size;
-        let block_stats = columns_statistics(block)?;
+        let block_stats = column_statistic::gen_columns_statistics(block)?;
         self.blocks_statistics.push(block_stats.clone());
         Ok(PartiallyAccumulated {
             accumulator: self,
@@ -146,146 +140,5 @@ impl PartiallyAccumulated {
         );
         stats.blocks_metas.push(block_meta);
         self.accumulator
-    }
-}
-
-pub struct BlockStatistics {
-    pub block_rows_size: u64,
-    pub block_bytes_size: u64,
-    pub block_file_location: String,
-    pub block_column_statistics: HashMap<ColumnId, ColumnStatistics>,
-    pub block_cluster_statistics: Option<ClusterStatistics>,
-}
-
-impl BlockStatistics {
-    pub fn from(
-        data_block: &DataBlock,
-        location: String,
-        cluster_stats: Option<ClusterStatistics>,
-    ) -> Result<BlockStatistics> {
-        Ok(BlockStatistics {
-            block_file_location: location,
-            block_rows_size: data_block.num_rows() as u64,
-            block_bytes_size: data_block.memory_size() as u64,
-            block_column_statistics: columns_statistics(data_block)?,
-            block_cluster_statistics: cluster_stats,
-        })
-    }
-
-    pub fn clusters_statistics(
-        cluster_key_id: u32,
-        cluster_key_index: &[usize],
-        block: &DataBlock,
-    ) -> Result<Option<ClusterStatistics>> {
-        if cluster_key_index.is_empty() {
-            return Ok(None);
-        }
-
-        let mut min = Vec::with_capacity(cluster_key_index.len());
-        let mut max = Vec::with_capacity(cluster_key_index.len());
-
-        for key in cluster_key_index.iter() {
-            let col = block.column(*key);
-
-            let mut left = col.get_checked(0)?;
-            // To avoid high cardinality, for the string column,
-            // cluster statistics uses only the first 5 bytes.
-            if let DataValue::String(v) = &left {
-                let l = v.len() as usize;
-                let e = if l < 5 { l } else { 5 };
-                left = DataValue::from(&v[0..e]);
-            }
-            min.push(left);
-
-            let mut right = col.get_checked(col.len() - 1)?;
-            if let DataValue::String(v) = &right {
-                let l = v.len() as usize;
-                let e = if l < 5 { l } else { 5 };
-                right = DataValue::from(&v[0..e]);
-            }
-            max.push(right);
-        }
-
-        Ok(Some(ClusterStatistics {
-            cluster_key_id,
-            min,
-            max,
-        }))
-    }
-}
-
-pub fn columns_statistics(data_block: &DataBlock) -> Result<StatisticsOfColumns> {
-    let mut statistics = StatisticsOfColumns::new();
-
-    let leaves = traverse::traverse_columns_dfs(data_block.columns())?;
-
-    for (idx, col) in leaves.iter().enumerate() {
-        let col_data_type = col.data_type();
-        if !MinMaxIndex::is_supported_type(&col_data_type) {
-            continue;
-        }
-
-        // later, during the evaluation of expressions, name of field does not matter
-        let data_field = DataField::new("", col_data_type);
-        let column_field = ColumnWithField::new(col.clone(), data_field);
-        let mut min = DataValue::Null;
-        let mut max = DataValue::Null;
-
-        let rows = col.len();
-
-        let mins = eval_aggr("min", vec![], &[column_field.clone()], rows)?;
-        let maxs = eval_aggr("max", vec![], &[column_field], rows)?;
-
-        if mins.len() > 0 {
-            min = mins.get(0);
-        }
-        if maxs.len() > 0 {
-            max = maxs.get(0);
-        }
-        let (is_all_null, bitmap) = col.validity();
-        let unset_bits = match (is_all_null, bitmap) {
-            (true, _) => rows,
-            (false, Some(bitmap)) => bitmap.unset_bits(),
-            (false, None) => 0,
-        };
-
-        let in_memory_size = col.memory_size() as u64;
-        let col_stats = ColumnStatistics {
-            min,
-            max,
-            null_count: unset_bits as u64,
-            in_memory_size,
-        };
-
-        statistics.insert(idx as u32, col_stats);
-    }
-    Ok(statistics)
-}
-
-mod traverse {
-    use super::*;
-
-    // traverses columns and collects the leaves in depth first manner
-    pub(crate) fn traverse_columns_dfs(columns: &[ColumnRef]) -> Result<Vec<ColumnRef>> {
-        let mut leaves = vec![];
-        for f in columns {
-            traverse_recursive(f, &mut leaves)?;
-        }
-        Ok(leaves)
-    }
-
-    fn traverse_recursive(column: &ColumnRef, leaves: &mut Vec<ColumnRef>) -> Result<()> {
-        match column.data_type() {
-            DataTypeImpl::Struct(_) => {
-                let struct_col: &StructColumn = Series::check_get(column)?;
-                for f in struct_col.values() {
-                    traverse_recursive(f, leaves)?
-                }
-            }
-            _ => {
-                leaves.push(column.clone());
-            }
-        }
-        Ok(())
     }
 }
