+++ query/tests/it/storages/fuse/statistics.rs
@@ -19,16 +19,17 @@ use common_datablocks::DataBlock;
 use common_datavalues::prelude::*;
 use common_fuse_meta::meta::ColumnStatistics;
 use databend_query::storages::fuse::statistics::accumulator;
+use databend_query::storages::fuse::statistics::gen_columns_statistics;
 use databend_query::storages::fuse::statistics::reducers;
+use databend_query::storages::fuse::statistics::BlockStatistics;
 
-use crate::storages::fuse::statistics::accumulator::BlockStatistics;
 use crate::storages::fuse::table_test_fixture::TestFixture;
 
 #[test]
 fn test_ft_stats_block_stats() -> common_exception::Result<()> {
     let schema = DataSchemaRefExt::create(vec![DataField::new("a", i32::to_data_type())]);
     let block = DataBlock::create(schema, vec![Series::from_data(vec![1, 2, 3])]);
-    let r = accumulator::columns_statistics(&block)?;
+    let r = gen_columns_statistics(&block)?;
     assert_eq!(1, r.len());
     let col_stats = r.get(&0).unwrap();
     assert_eq!(col_stats.min, DataValue::Int64(1));
@@ -45,7 +46,7 @@ fn test_ft_stats_col_stats_reduce() -> common_exception::Result<()> {
     let blocks = TestFixture::gen_sample_blocks_ex(num_of_blocks, rows_per_block, val_start_with);
     let col_stats = blocks
         .iter()
-        .map(|b| accumulator::columns_statistics(&b.clone().unwrap()))
+        .map(|b| gen_columns_statistics(&b.clone().unwrap()))
         .collect::<common_exception::Result<Vec<_>>>()?;
     let r = reducers::reduce_block_statistics(&col_stats);
     assert!(r.is_ok());
